backend/requirements.txt
# Python 3.13 compatible versions
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
anthropic==0.47.0
pytesseract==0.3.13
pillow==11.1.0
celery==5.4.0
redis==5.2.1
caldav==1.3.9
aiofiles==24.1.0
